pyproject.toml
[project]
name = "harsh-engine"
version = "0.1.0"
description = "A simple web app intentionally designed to have numerous common vulnerabilities."
authors = [
    {name = "Noah Grattan"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
